exercise_2/backend/requirements.txt
fastapi>=0.112,<1
uvicorn[standard]>=0.30,<1
python-dotenv>=1.0,<2
openai>=1.35,<2
httpx>=0.25,<1


